Break out cocktails components into components folder
